+++ web/components/General/Icons.tsx
@@ -265,3 +265,20 @@ export const ArrowLeftIcon = ({ size = 20, className = "" }: IconProps) => (
     <path d="M19 12H5" />
   </svg>
 );
+
+export const UserIcon = ({ size = 24, className = "" }: IconProps) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className={className}
+  >
+    <path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2" />
+    <circle cx="12" cy="7" r="4" />
+  </svg>
+);
